requirements.txt
streamlit
transformers
torch
Pillow
scikit-learn
pandas
spacy
matplotlib
sastrawi
nltk
joblib
huggingface_hub
imblearn
seaborn
openpyxl
